requirements.txt
git+https://github.com/cthoyt/p_tqdm@patch-1
pymongo
qwikidata
sklearn
tqdm
attrs
click
cssselect
dask
dictances
editdistance
lxml
matplotlib
numpy
orjson
pandas
requests
scipy
unicodeblock
csvkit
pgmpy
flyingsquid
black
mypy
flake8
ipython
more-itertools
datasketch
rich
